TeamCity: Consolidate RSA crypto into login scanner
